+++ src/BudgetBakers.Wallet.Net/Dtos/AmountWithCurrencyDto.cs
@@ -8,6 +8,6 @@ namespace BudgetBakers.Wallet.Net.Dtos
         public string? CurrencyCode { get; set; }
 
         [JsonProperty("value")]
-        public double? Value { get; set; }
+        public decimal? Value { get; set; }
     }
 }
