+++ src/BudgetBakers.Wallet.Net/Models/Budget/LimitOverride.cs
@@ -3,7 +3,7 @@ namespace BudgetBakers.Wallet.Net.Models.Budget
     public class LimitOverride
     {
         public required string Period { get; set; }
-        public double? Limit { get; set; }
+        public decimal? Limit { get; set; }
         public bool? SetBaseline { get; set; }
     }
 }
